+++ lua/obsidian/commands/init.lua
@@ -231,6 +231,8 @@ M.register("open", { nargs = "*", complete = M.note_complete })
 
 M.register("tags", { nargs = "*" })
 
+M.register("rename_tag", { nargs = "*" })
+
 M.register("search", { nargs = "?" })
 
 M.register("sync", {